Add variable declaration statements
Added variable declaration statements to the parser; note that they
don't execute and will currently cause the interpreter to panic.

Added tests for the variable declaration statements. Found and fixed a
small bug in the helper function for creating an ident token.

Changed `ParseError::ExpectedToken` variant to have an expected
`TokenType` and a found `Token`, and updated the error handling so that
it points out the found `Token`.

Added a match_token helper function to the parser.