Multiple error handling in parser.
Starts to address #5. Error handling is not yet what it should be, but
multiple errors are handled in the parser. The `synchronize` method
doesn't work quite as expected because the token where a semicolon
should be gets consumed, so something like

```
print 8 // oops
print 8; // okay
```
Doesn't output anything because the second `print` doesn't get shown to
the synchronizer.