Change the `LoxValue::Callable` variant to have Option Environment
I was having a problem where calling `dbg!(&env)` after the environments
were populated with builtin functions, where I had an infinite loop, and
this was making debugging challenging. I fixed this by making the
`Callable` variant of `LoxValue` point to an
`Option<Rc<RefCell<Environment>>>`. Then for defining the builtins, I
set this enum to `None`, so that the builtins would not point back to
the `Interpreter`s `globals`. Instead, in the `Callable::call`
implementation, it checks this `Option`, and if the variant is `None`,
it points back to the interpreter globals. A bit of a workaround, but
hopefully this will help to debug the rest of chapter 10.